requirements.txt
# Core Flask packages
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.1.1
WTForms==3.0.1

# ML and NLP packages
tensorflow==2.13.0
nltk==3.8.1
scikit-learn==1.3.0
numpy==1.24.3
pandas==1.5.3
transformers==4.33.0
torch==2.0.1

# WhatsApp/SMS Integration
twilio==8.5.0
Flask-SocketIO==5.3.6

# Translation and Multilingual Support  
google-cloud-translate==3.12.0
langdetect==1.0.9

# Government API and Data Processing
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
celery==5.3.1
redis==4.6.0

# Additional utilities
Pillow==10.0.0
python-dateutil==2.8.2
pytz==2023.3
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.7

# Caching and Performance
Flask-Caching==2.1.0

# API Documentation
Flask-RESTX==1.1.0

# Testing
pytest==7.4.0